# Optimize CastHooks::castStringToDate/Timestamp to avoid throwing exceptions