… velox/functions/lib/tests/DateTimeFormatterTest.cpp
… velox/functions/prestosql/tests/SequenceTest.cpp
… velox/functions/sparksql/tests/DateTimeFunctionsTest.cpp
… velox/functions/sparksql/tests/MakeTimestampTest.cpp
… velox/type/tests/TimestampConversionTest.cpp