Make casts from VARCHAR to DATE and TIMESTAMP non-throwing.

Convert 'from_iso8601_date' Presto function to non-throwing.

These changes improve performance of casts from empty or invalid strings by
100x. Note that cast from empty string to TIMESTAMP has been optimized already,
hence, these changes do not improve that use case.

Before:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast_varhar_as_date##try_cast_invalid_empty_inp             12.42s    80.49m
cast_varhar_as_date##tryexpr_cast_invalid_empty             16.66s    60.01m
cast_varhar_as_date##try_cast_invalid_input                 12.43s    80.46m
cast_varhar_as_date##tryexpr_cast_invalid_input             16.86s    59.30m
cast_varhar_as_date##cast_valid                            35.18ms     28.43
----------------------------------------------------------------------------
cast_varhar_as_timestamp##try_cast_invalid_empt            26.31ms     38.00
cast_varhar_as_timestamp##tryexpr_cast_invalid_            54.89ms     18.22
cast_varhar_as_timestamp##try_cast_invalid_inpu              6.99s   143.12m
cast_varhar_as_timestamp##tryexpr_cast_invalid_              9.00s   111.17m
cast_varhar_as_timestamp##cast_valid                       43.74ms     22.86
----------------------------------------------------------------------------
```

After:

```
cast_varhar_as_date##try_cast_invalid_empty_inp            49.96ms     20.02
cast_varhar_as_date##tryexpr_cast_invalid_empty            83.63ms     11.96
cast_varhar_as_date##try_cast_invalid_input                59.09ms     16.92
cast_varhar_as_date##tryexpr_cast_invalid_input            96.01ms     10.42
cast_varhar_as_date##cast_valid                            35.89ms     27.86
----------------------------------------------------------------------------
cast_varhar_as_timestamp##try_cast_invalid_empt            25.88ms     38.64
cast_varhar_as_timestamp##tryexpr_cast_invalid_            51.45ms     19.43
cast_varhar_as_timestamp##try_cast_invalid_inpu            80.67ms     12.40
cast_varhar_as_timestamp##tryexpr_cast_invalid_           109.51ms      9.13
cast_varhar_as_timestamp##cast_valid                       49.48ms     20.21
----------------------------------------------------------------------------
```

Fixes https://github.com/facebookincubator/velox/issues/9914

Differential Revision: D57733757